+++ app/src/AppBundle/API/Restaurant/MealController.php
@@ -22,7 +22,7 @@ class MealController extends ApiBaseController
      * @param Request $request
      *
      *
-     * @REST\Post("/restaurant/{id}/tab/{idTab}/meal/create", name="api_create_meal")
+     * @REST\Post("/restaurants/{id}/tabs/{idTab}/meals/create", name="api_create_meal")
      * @REST\RequestParam(name="name")
      * @REST\RequestParam(name="description", nullable=true)
      * @REST\RequestParam(name="price", nullable=true)
@@ -35,7 +35,25 @@ class MealController extends ApiBaseController
     public function createMeal(Request $request, ParamFetcher $paramFetcher)
     {
         $user = $this->container->get('security.token_storage')->getToken()->getUser();
-        $restaurant = $this->getRestaurantRepository()->findOneBy(array("id" => $request->get('id')));
+
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
+
+        if (!$request->get('idTab')) {
+            return $this->helper->error('idTab', true);
+        } elseif (!preg_match('/\d/', $request->get('idTab'))) {
+            return $this->helper->error('param \'idTab\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
         $restaurantUsers = $restaurant->getUsers();
 
         if(!$this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN') &&
@@ -45,7 +63,10 @@ class MealController extends ApiBaseController
 
         $params = $paramFetcher->all();
 
-        $tab = $this->getTabMealRepository()->find($request->get('idTab'));
+        $tab = $elasticaManager->getRepository('AppBundle:TabMeal')->findById($request->get('idTab'));
+        if (!$tab) {
+            return $this->helper->elementNotFound('TabMeal');
+        }
 
         $meal = new Content();
         $meal->setStatus(1);
@@ -70,34 +91,54 @@ class MealController extends ApiBaseController
 
     /**
      *
-     * @REST\Put("/restaurant/{id}/meal/{idMeal}/edit", name="api_edit_meal")
+     * @REST\Put("/restaurants/{id}/meals/{idMeal}", name="api_edit_meal")
      *
      */
     public function editMeal(Request $request)
     {
         $user = $this->container->get('security.token_storage')->getToken()->getUser();
-        $restaurant = $this->getRestaurantRepository()->findOneBy(array("id" => $request->get('id')));
+
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
+
+        if (!$request->get('idMeal')) {
+            return $this->helper->error('idMeal', true);
+        } elseif (!preg_match('/\d/', $request->get('idMeal'))) {
+            return $this->helper->error('param \'idMeal\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
         $restaurantUsers = $restaurant->getUsers();
 
         if(!$this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN') &&
             !$restaurantUsers->contains($user)){
             return $this->helper->error('Vous n\'êtes pas autorisé à effectuer cette action');
         }
-
-        $meal=$this->getContentRepository()->findOneBy(array("id" => $request->get('idMeal')));
-
+        $meal = $elasticaManager->getRepository('AppBundle:Content')->findById($request->get('idMeal'));
+        if (!$meal) {
+            return $this->helper->elementNotFound('Meal');
+        }
 
         $request_data = $request->request->all();
 
-        if($request_data['name'] != null){
+        if($request_data['name']){
             $meal->setName($request_data['name']);
         }
-        if($request_data['price'] != null){
+        if($request_data['price']){
             $meal->setPrice($request_data['price']);
         }
-        if($request_data['description'] != null){
+        if($request_data['description']){
             $meal->setDescription($request_data['description']);
         }
+
         $em = $this->getEntityManager();
         $em->persist($meal);
 
@@ -106,24 +147,52 @@ class MealController extends ApiBaseController
 
     /**
      *
-     * @REST\Post("/restaurant/{id}/meal/{idMeal}", name="api_daily_stock")
+     * @REST\Put("/restaurants/{id}/meals/{idMeal}/daily_stock", name="api_daily_stock")
      * @REST\RequestParam(name="initialStock")
      */
-    public function updateDailyStock(Request $request,ParamFetcher $paramFetcher)
+    public function updateDailyStock(Request $request, ParamFetcher $paramFetcher)
     {
         $user = $this->container->get('security.token_storage')->getToken()->getUser();
-        $restaurant = $this->getRestaurantRepository()->findOneBy(array("id" => $request->get('id')));
+
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
+
+        if (!$request->get('idMeal')) {
+            return $this->helper->error('idMeal', true);
+        } elseif (!preg_match('/\d/', $request->get('idMeal'))) {
+            return $this->helper->error('param \'idMeal\' must be an integer');
+        }
+
+        if (!$request->get('initialStock')) {
+            return $this->helper->error('initialStock', true);
+        } elseif (!preg_match('/\d/', $request->get('initialStock'))) {
+            return $this->helper->error('param \'initialStock\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
         $restaurantUsers = $restaurant->getUsers();
 
         if(!$this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN') &&
             !$restaurantUsers->contains($user)){
             return $this->helper->error('Vous n\'êtes pas autorisé à effectuer cette action');
         }
-        $params=$paramFetcher->all();
-        $meal = $this->getContentRepository()->findOneBy(array('id' => $request->get('idMeal'), 'restaurant' => $restaurant, 'status'=> true));
 
-        $meal->setInitialStock($params['initialStock']);
-        $meal->setCurrentStock($params['initialStock']);
+        $params=$paramFetcher->all();
+        $meal = $elasticaManager->getRepository('AppBundle:Content')->findById($request->get('idMeal'));
+        if (!$meal) {
+            return $this->helper->elementNotFound('Meal');
+        }
+
+        $meal->setInitialStock($meal->getCurrentStock() + $params['initialStock']);
+        $meal->setCurrentStock($meal->getInitialStock());
         $em = $this->getEntityManager();
         $em->persist($meal);
         $em->flush();
@@ -133,21 +202,43 @@ class MealController extends ApiBaseController
 
     /**
      *
-     * @REST\Post("/restaurant/{id}/meal/{idMeal}/stock", name="api_stock_change")
+     * @REST\Put("/restaurants/{id}/meals/{idMeal}/update_stock", name="api_stock_change")
      * @REST\RequestParam(name="stock")
      */
-    public function setCurrentStock(Request $request,ParamFetcher $paramFetcher)
+    public function updateCurrentStock(Request $request,ParamFetcher $paramFetcher)
     {
         $user = $this->container->get('security.token_storage')->getToken()->getUser();
-        $restaurant = $this->getRestaurantRepository()->findOneBy(array("id" => $request->get('id')));
+
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
+
+        if (!$request->get('idMeal')) {
+            return $this->helper->error('idMeal', true);
+        } elseif (!preg_match('/\d/', $request->get('idMeal'))) {
+            return $this->helper->error('param \'idMeal\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
         $restaurantUsers = $restaurant->getUsers();
 
         if(!$this->get('security.authorization_checker')->isGranted('ROLE_SUPER_ADMIN') &&
             !$restaurantUsers->contains($user)){
             return $this->helper->error('Vous n\'êtes pas autorisé à effectuer cette action');
         }
+
         $params=$paramFetcher->all();
-        $meal = $this->getContentRepository()->findOneBy(array('id' => $request->get('idMeal'), 'restaurant' => $restaurant, 'status'=> true));
+        $meal = $elasticaManager->getRepository('AppBundle:Content')->findById($request->get('idMeal'));
+        if (!$meal) {
+            return $this->helper->elementNotFound('Meal');
+        }
 
         $meal->setCurrentStock($meal->getCurrentStock() + $params['stock']);
 
@@ -167,25 +258,59 @@ class MealController extends ApiBaseController
      */
     public function getMeal(Request $request)
     {
-        $restaurant = $this->getRestaurantRepository()->find($request->get('id'));
-        $meal = $this->getContentRepository()->findBy(array('id' => $request->get('idMeal'), 'restaurant' => $restaurant, 'status'=> true));
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
 
+        if (!$request->get('idMeal')) {
+            return $this->helper->error('idMeal', true);
+        } elseif (!preg_match('/\d/', $request->get('idMeal'))) {
+            return $this->helper->error('param \'idMeal\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
+        $meal = $elasticaManager->getRepository('AppBundle:Content')->findById($request->get('idMeal'));
+        if (!$meal) {
+            return $this->helper->elementNotFound('Meal');
+        }
 
         return $this->helper->success($meal, 200);
     }
-
+  
     /**
-     * @REST\Delete("/restaurant/{id}/meal/{idMeal}", name="api_delete_meal")
+     * @REST\Delete("/restaurants/{id}/meals/{idMeal}", name="api_delete_meal")
      */
     public function deleteMeal(Request $request)
     {
-        $restaurant = $this->getRestaurantRepository()->find($request->get('id'));
-        $meal = $this->getContentRepository()->findOneBy(
-            array(
-                'status' => true,
-                'restaurant' => $restaurant,
-                'id' => $request->get('idMeal')
-            ));
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
+
+        if (!$request->get('idMeal')) {
+            return $this->helper->error('idMeal', true);
+        } elseif (!preg_match('/\d/', $request->get('idMeal'))) {
+            return $this->helper->error('param \'idMeal\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
+        $meal = $elasticaManager->getRepository('AppBundle:Content')->findById($request->get('idMeal'));
+        if (!$meal) {
+            return $this->helper->elementNotFound('Meal');
+        }
 
         $em = $this->getEntityManager();
         $em->remove($meal);
@@ -196,31 +321,59 @@ class MealController extends ApiBaseController
 
     /**
      *
-     * @REST\Get("/restaurant/{id}/meals", name="api_list_meals")
+     * @REST\Get("/restaurants/{id}/meals", name="api_list_meals")
      *
      */
     public function getMeals(Request $request)
     {
-        $restaurant = $this->getRestaurantRepository()->find($request->get('id'));
-        $meals = $this->getContentRepository()->findBy(array('restaurant' => $restaurant,'type' => Content::TYPE_MEAL));
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
+        $meals = $elasticaManager->getRepository('AppBundle:Content')->findByRestaurant($restaurant, Content::TYPE_MEAL);
+
         return $this->helper->success($meals, 200);
     }
 
     /**
      *
-     * @REST\Get("/restaurant/{id}/tab/{idTab}/meals", name="api_list_meals_by_tab")
+     * @REST\Get("/restaurants/{id}/tabs/{idTab}/meals", name="api_list_meals_by_tab")
      *
      */
     public function getMealsFromTab(Request $request)
     {
-        $restaurant = $this->getRestaurantRepository()->find($request->get('id'));
-        $tab = $this->getTabMealRepository()->find($request->get('idTab'));
+        if (!$request->get('id')) {
+            return $this->helper->error('id', true);
+        } elseif (!preg_match('/\d/', $request->get('id'))) {
+            return $this->helper->error('param \'id\' must be an integer');
+        }
 
-        $meals = $this->getContentRepository()->findBy(array(
-            'restaurant' => $restaurant,
-            'tab' => $tab,
-            'type' => Content::TYPE_MEAL
-        ));
+        if (!$request->get('idTab')) {
+            return $this->helper->error('idTab', true);
+        } elseif (!preg_match('/\d/', $request->get('idTab'))) {
+            return $this->helper->error('param \'idTab\' must be an integer');
+        }
+
+        $elasticaManager = $this->container->get('fos_elastica.manager');
+        $restaurant = $elasticaManager->getRepository('AppBundle:Restaurant')->findById($request->get('id'));
+        if (!$restaurant) {
+            return $this->helper->elementNotFound('Restaurant');
+        }
+
+        $tab = $elasticaManager->getRepository('AppBundle:TabMeal')->findById($request->get('idTab'));
+        if (!$tab) {
+            return $this->helper->elementNotFound('TabMeal');
+        }
+
+        $meals = $elasticaManager->getRepository('AppBundle:Content')->findByTab($tab, Content::TYPE_MEAL);
 
         return $this->helper->success($meals, 200);
     }
